Avoid unnecessary cast
warning: casting to the same type is unnecessary (`u64` -> `u64`)
 --> benches/reversedbitreader_bench.rs:8:16
  |
8 |         sum += br.get_bits(*x).unwrap() as u64;
  |                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `br.get_bits(*x).unwrap()`
  |
  = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
  = note: `#[warn(clippy::unnecessary_cast)]` on by default

+++ benches/reversedbitreader_bench.rs
@@ -5,7 +5,7 @@ use ruzstd::decoding::bit_reader_reverse::BitReaderReversed;
 fn do_all_accesses(br: &mut BitReaderReversed, accesses: &[u8]) -> u64 {
     let mut sum = 0;
     for x in accesses {
-        sum += br.get_bits(*x).unwrap() as u64;
+        sum += br.get_bits(*x).unwrap();
     }
     let _ = black_box(br);
     sum
